carve out component for architecture
